+++ PiSynth/Patch.h
@@ -15,9 +15,9 @@ private:
 	std::map<int, float> outputLevels;
 	std::map<int, float> modSens;
 	//ENVELOPE GENERATOR DATA
-	std::map<int, std::vector<int>> expCoeffs;
+	std::map<int, std::vector<float>> expCoeffs;
 	std::map<int, std::vector<float>> targetVals;
-	std::map<int, std::vector<int>> holdTimes;
+	std::map<int, std::vector<float>> holdTimes;
 
 public:
 	Patch();
@@ -32,7 +32,7 @@ public:
 	void setOutputLevel(int operatorID, float outputLevel);
 	void setModulationSensitivity(int operatorID, float modulationSensitivity);
 	//ENVELOPE GENERATOR DATA
-	void addEnvelopeSegment(int operatorID, int speed, float targetVal, int holdTime);
+	void addEnvelopeSegment(int operatorID, float speed, float targetVal, float holdTime);
 
 	//-------------------- GETTERS ------------------------//
 	//ALGORITHM DATA
@@ -44,9 +44,9 @@ public:
 	float getModulationSensitivity(int operatorID);
 	//ENVELOPE GENERATOR DATA
 	int getEnvSegmentCount(int operatorID);
-	std::vector<int> getCoefficients(int operatorID);
+	std::vector<float> getCoefficients(int operatorID);
 	std::vector<float> getTargets(int operatorID);
-	std::vector<int> getHoldTimes(int operatorID);
+	std::vector<float> getHoldTimes(int operatorID);
 
 	//DEBUG
 	void printPatchData();
